revert: 回退 Phase P2 体验打磨变更，准备重新设计交互方案
回退 74d5760 和 4a045a1，保留工程修复（P2-3 pass_threshold / P2-4 chat_json）
待重新设计后重新实现 P2-1（批测配置）和 P2-2（配置展示）

+++ backend/app/services/judge_runner.py
@@ -1,5 +1,6 @@
 import json
 import logging
+from decimal import Decimal
 
 from app.llm.base import LLMAdapter
 from app.models.judge_config import ChecklistItem, EvalDimension
@@ -28,7 +29,7 @@ class JudgeRunner:
         system_prompt: str,
         checklist_items: list[ChecklistItem],
         eval_dimensions: list[EvalDimension],
-        pass_threshold: float,
+        pass_threshold: Decimal,
         temperature: float | None = None,
         max_tokens: int | None = None,
     ):
@@ -166,7 +167,7 @@ class JudgeRunner:
         # Evaluation average >= threshold
         if eval_scores:
             avg = sum(s["score"] for s in eval_scores) / len(eval_scores)
-            if avg < self.pass_threshold:
+            if avg < float(self.pass_threshold):
                 return False
 
         return True
